perf(int3): fp16 VQ codebook — MoE gpu 228→194ms, 2.56→2.85 tok/s
The VQ MoE dot is gather-latency-bound: each subvector does a random idx→cb[idx]
lookup then 4 f32 MACs. At f32 the codebook is VQ_K·VQ_DIM·4 = 64KB — L2-resident,
too big for the 32KB L1. Narrowing it to fp16 (32KB) fits L1, so the random gather
becomes an L1 hit, and each gather is 8B instead of 16B.

Quality: measured bf16 vs fp16 codebook error against the exact CPU reference
(quant_vq + matvec_vq) — bf16's 8-bit mantissa FAILS the 1e-3 kernel-oracle tol at
these reduction lengths (rel ~1.7e-3), fp16's 10-bit mantissa clears it with ~5.6×
margin (rel ~2e-4). fp16 is the same 2 bytes as bf16, so identical cache/bandwidth
win with strictly better accuracy. Activations stay f32 (the gather, not the x
stream, is the bottleneck) to keep the margin.

- math::f32_to_f16: host f32→fp16 (RNE, subnormals, inf/nan), self-checked.
- dot_vq_wave: codebook const __half*, 8B two-__half2 gather, products in f32.
- moe.hip / linalg.hip / hip.rs / gpu.rs / pin.rs: codebook f32→u16 (fp16) end to
  end; upload narrows f32→fp16 at load (also halves the resident codebook footprint).
- kernel oracle uploads the codebook fp16 while the reference keeps f32, so gemv_vq
  + moe_vq now validate the fp16 rounding directly. All 7 kernel oracles green on GPU.

Measured (128-tok bench, --attn dense): MoE gpu 228→194ms (−15%), wall 391→350ms,
2.56→2.85 tok/s (+11%), output coherent, expert hit 83%.

+++ src/gpu.rs
@@ -226,7 +226,8 @@ pub struct GpuEngine<'a> {
     // Per-token host build scratch — reused every layer so the hot path allocates
     // nothing: resolved VQ descriptors + weights, the resolved batch, D2H staging.
     w: Vec<f32>,
-    codebooks: [*const f32; 3],
+    /// The three per-projection VQ codebooks (gate/up/down), fp16, resident.
+    codebooks: [*const u16; 3],
     mlps_vq: Vec<MlpVq>,
     descs_vq: Vec<ExpertDescVq>,
     gl_host: Vec<u8>,
